+++ models/BaseMetaModel.php
@@ -12,6 +12,8 @@
 
 namespace vistart\Models\models;
 
+use vistart\Models\traits\MetaTrait;
+
 /**
  * Description of BaseMetaModel
  *
@@ -20,9 +22,11 @@ namespace vistart\Models\models;
  */
 abstract class BaseMetaModel extends BaseBlameableModel
 {
+    use MetaTrait;
 
     public $idAttribute = 'key';
     public $idPreassigned = true;
+    public $idAttributeLength = 255;
     public $createdAtAttribute = false;
     public $updatedAtAttribute = false;
     public $enableIP = false;
@@ -30,44 +34,4 @@ abstract class BaseMetaModel extends BaseBlameableModel
     public $updatedByAttribute = false;
     public $confirmationAttribute = false;
 
-    /**
-     * Store the guid of blame.
-     * @var string 
-     */
-    public $blameGuid = '';
-
-    public function behaviors()
-    {
-        return $this->getMetaBehaviors();
-    }
-
-    /**
-     * Skip all behaviors of parent class.
-     * @return array
-     */
-    public function getMetaBehaviors()
-    {
-        return [];
-    }
-
-    public function init()
-    {
-        $this->on(static::EVENT_INIT, [$this, 'onInitBlameGuid']);
-        parent::init();
-    }
-
-    /**
-     * Initialize blame's guid.
-     * @param \yii\base\Event $event
-     * @throws \yii\base\InvalidConfigException
-     */
-    public function onInitBlameGuid($event)
-    {
-        $sender = $event->sender;
-        if (empty($sender->blameGuid)) {
-            throw new \yii\base\InvalidConfigException('Empty blame guid is not allowed.');
-        }
-        $createdByAttribute = $sender->createdByAttribute;
-        $sender->$createdByAttribute = $sender->blameGuid;
-    }
 }